+++ docs/source/changelog.rst
@@ -16,6 +16,7 @@ Version 0.3.1 (2023-03-28)
 
 * Fixed issue for path based query, where Rickle objects are considered callable (rightfully).
 * Uses `inspect.isfunction` instead of `callable`.
+* Added `meta` to base class for getting metadata of a property.
 
 Version 0.3.0 (2023-02-16)
 --------------------------